api 모델 변수 형식 수정

+++ app/src/main/java/com/example/wtw_weather_app/api_model/CurrentAndHourly.kt
@@ -4,9 +4,9 @@ data class CurrentAndHourly(
     var dt: Long,
     var temp: Double,
     var feels_like: Double,
-    var humidity: Double,
+    var humidity: Int,
     var clouds: Int,
-    var rain: Double,
-    var snow: Double,
+    var rain: Int,
+    var snow: Int,
     var weather: ArrayList<Weather>
 )
